refactor: update component categorization from Atoms to Molecules in stories
feat: add ShortcutBadge component for keyboard shortcut indicators

feat: introduce VisibilityBadge component for metric visibility tiers

docs: consolidate and document all design-system building blocks in Primitives story

fix: adjust layout parameters and remove unused tags in various stories
