+++ docs/atomic-design-remediation-plan.md
@@ -304,7 +304,7 @@ The 30+ design system primitives in `playground/src/components/ui/` (alert, avat
 
 ---
 
-## Phase 4 — Source Component Refactors (anti-pattern fixes)
+## Phase 4 — Source Component Refactors (anti-pattern fixes) ✅ COMPLETE
 
 **Goal:** Fix the 4 anti-patterns identified in the classification audit. These are source code changes, not Storybook changes.
