+++ src/components/list/DefaultListItem.tsx
@@ -1,24 +1,10 @@
 import { cn } from '@/lib/utils';
 import type { IListItem, ListItemContext } from './types';
+import { ShortcutBadge } from './ShortcutBadge';
 
 interface DefaultListItemProps<TPayload> {
   item: IListItem<TPayload>;
   ctx: ListItemContext;
-}
-
-function ShortcutBadge({ tokens }: { tokens: string[] }) {
-  return (
-    <span className="flex items-center gap-0.5 ml-auto shrink-0">
-      {tokens.map((token, i) => (
-        <kbd
-          key={i}
-          className="inline-flex items-center rounded border border-zinc-300 bg-zinc-100 px-1 py-0.5 text-[10px] font-mono text-zinc-500 dark:border-zinc-600 dark:bg-zinc-800 dark:text-zinc-400"
-        >
-          {token === 'meta' ? '⌘' : token === 'shift' ? '⇧' : token === 'alt' ? '⌥' : token}
-        </kbd>
-      ))}
-    </span>
-  );
 }
 
 export function DefaultListItem<TPayload>({ item, ctx }: DefaultListItemProps<TPayload>) {
