--- stories/catalog/atoms/DropdownMenu.stories.tsx
@@ -1,44 +1,0 @@
-/**
- * Catalog / Atoms / DropdownMenu
- *
- * Radix-based dropdown menu with labels, items, and separators.
- */
-
-import React from 'react';
-import type { Meta, StoryObj } from '@storybook/react';
-import { Button } from '@/components/ui/button';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-
-const meta: Meta = {
-  title: 'catalog/atoms/interactive/DropdownMenu',
-  parameters: { layout: 'centered' },
-};
-export default meta;
-type Story = StoryObj;
-
-export const Default: Story = {
-  name: 'DropdownMenu',
-  render: () => (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="outline">Open Dropdown</Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent>
-        <DropdownMenuLabel>Actions</DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem>View details</DropdownMenuItem>
-        <DropdownMenuItem>Edit entry</DropdownMenuItem>
-        <DropdownMenuItem>Clone to today</DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem className="text-destructive">Delete</DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  ),
-};
