+++ stories/catalog/molecules/CommandInput.stories.tsx
@@ -1,8 +1,13 @@
 /**
- * Catalog / Atoms / CommandInput
+ * Catalog / Molecules / CommandInput
  *
  * The search-input row used at the top of every command-palette variant.
  * Consists of: leading Search icon · text input · trailing Esc badge.
+ *
+ * ⚠️  Catalog surrogate: `CommandInputRow` here is a standalone demo component.
+ *     In production, `CommandPalette.tsx` uses `cmdk`'s `Command.Input` directly
+ *     (from the `cmdk` library), which is tightly integrated with `Command.List`
+ *     for accessible search. This story documents the visual design pattern only.
  *
  * Stories:
  *  1. Empty        – no value, default placeholder
